criei o Consiousness view
